Refactor: Targeting the latest version of android and more
-Moving all dependencies to the latest version
-Supporting android 11
-Removing all depreciated code and using the latest specifications of MVVM
