+++ .idea/jarRepositories.xml
@@ -1,0 +1,40 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="RemoteRepositoriesConfiguration">
+    <remote-repository>
+      <option name="id" value="central" />
+      <option name="name" value="Maven Central repository" />
+      <option name="url" value="https://repo1.maven.org/maven2" />
+    </remote-repository>
+    <remote-repository>
+      <option name="id" value="jboss.community" />
+      <option name="name" value="JBoss Community repository" />
+      <option name="url" value="https://repository.jboss.org/nexus/content/repositories/public/" />
+    </remote-repository>
+    <remote-repository>
+      <option name="id" value="BintrayJCenter" />
+      <option name="name" value="BintrayJCenter" />
+      <option name="url" value="https://jcenter.bintray.com/" />
+    </remote-repository>
+    <remote-repository>
+      <option name="id" value="Google" />
+      <option name="name" value="Google" />
+      <option name="url" value="https://dl.google.com/dl/android/maven2/" />
+    </remote-repository>
+    <remote-repository>
+      <option name="id" value="D:\Android-SDK\extras\m2repository" />
+      <option name="name" value="D:\Android-SDK\extras\m2repository" />
+      <option name="url" value="file:/D:/Android-SDK/extras/m2repository/" />
+    </remote-repository>
+    <remote-repository>
+      <option name="id" value="D:\Android-SDK\extras\android\m2repository" />
+      <option name="name" value="D:\Android-SDK\extras\android\m2repository" />
+      <option name="url" value="file:/D:/Android-SDK/extras/android/m2repository/" />
+    </remote-repository>
+    <remote-repository>
+      <option name="id" value="D:\Android-SDK\extras\google\m2repository" />
+      <option name="name" value="D:\Android-SDK\extras\google\m2repository" />
+      <option name="url" value="file:/D:/Android-SDK/extras/google/m2repository/" />
+    </remote-repository>
+  </component>
+</project>
