+++ app/src/androidTest/java/com/example/umangburman/databindingwithlivedata/ExampleInstrumentedTest.java
@@ -1,8 +1,8 @@
 package com.example.umangburman.databindingwithlivedata;
 
 import android.content.Context;
-import android.support.test.InstrumentationRegistry;
-import android.support.test.runner.AndroidJUnit4;
+import androidx.test.platform.app.InstrumentationRegistry;
+import androidx.test.ext.junit.runners.AndroidJUnit4;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
